Update 0.0.3
- Implemented matrix stack
- Optimized shaders
- Added FXAA antialiasing
- Updated demo scene
- Bug fixes and refactoring

+++ shaders/fs_grayscale.glsl
@@ -1,4 +1,4 @@
-precision highp float;
+precision mediump float;
 
 varying vec2 vVertexPosition;
 uniform sampler2D uTex;
@@ -8,7 +8,7 @@ uniform float vz;
 
 void main()
 {
-    vec2 texCoord = vec2( vVertexPosition.s, vVertexPosition.t ) * 0.5 + 0.5;
+    vec2 texCoord = vec2(vVertexPosition.s, vVertexPosition.t) * 0.5 + 0.5;
     vec4 color = texture2D(uTex, texCoord);
     float gray = dot(color.rgb, vec3(vx, vy, vz));
     gl_FragColor = vec4(vec3(gray), 1.0);
